+++ hpc/nvidia/cuda/solver/solver.cu
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: GPL-3.0
+/* Copyright (c) 2025 Rong Tao */
+#include <stdio.h>
+#include "cuda_compat.h"
+#include "cuda_helpers.h"
+
+
+int main(int argc, char *argv[])
+{
+	int version = 0;
+
+	cusolverGetVersion(&version);
+	printf("Solver version %d\n", version);
+
+	return 0;
+}